% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jobs.R
\name{bg_func}
\alias{bg_func}
\title{Background parallel processing of jobs using purrr and tryCatch}
\usage{
bg_func(jobs, n_daemons)
}
\arguments{
\item{jobs}{A list of job specifications, each containing
a function (\code{fun}) and arguments (\code{args}).}

\item{n_daemons}{Number of parallel workers}
}
\value{
A list of results, with error messages in case of failure.
}
\description{
Executes a list of job specifications in parallel,
applies the given function with error handling,
and collects the results or error messages.
}
