% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jobs.R
\name{run_bg}
\alias{run_bg}
\title{Generic for running background jobs in bakerrr}
\usage{
run_bg(x, ...)
}
\arguments{
\item{x}{A \code{bakerrr} S7 job object.}

\item{...}{Not used. For future expansion.}
}
\value{
The input \code{x}, invisibly, after launching background jobs.
}
\description{
Initiates background execution for a bakerrr job object,
launching jobs via \code{callr::r_bg}.
Stores process status in \code{bg_job_status}.
}
