% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Status method for bakerrr S7 job objects}
\usage{
status(x, ...)
}
\arguments{
\item{x}{A \code{bakerrr} S7 job object.}

\item{...}{Further arguments (ignored).}
}
\value{
One of "waiting", "running", "done".
}
\description{
Returns job status as "waiting", "running", or "done".
}
